added setup commands
