+++ packages/anon-aadhaar-pcd/package.json
@@ -7,18 +7,12 @@
   "license": "MIT",
   "scripts": {
     "build": "tsup src/index.ts",
-    "circom": "circom",
-    "dev-setup": "/bin/bash ./script/setup-dev.sh setup",
-    "dev-install": "/bin/bash ./script/setup-dev.sh install",
     "test": "npx ts-mocha test/**/*.test.ts --exit",
     "lint": "eslint --ext .js,.ts .",
     "format": "prettier --ignore-path .prettierignore --write \"**/*.+(js|ts|json)\"",
     "prepare": "cd ../.. && husky install",
-    "gen-pdf": "/bin/bash ./script/setup-dev.sh pdf-setup",
     "test-publish": "yarn build && yarn lint && yarn test && npm publish --dry-run",
-    "publishtonpm": "yarn build && yarn lint && yarn test && npm publish",
-    "gen-contract": "/bin/bash ./script/setup-dev.sh contract-setup",
-    "gen-proof": "/bin/bash ./script/setup-dev.sh gen-proof"
+    "publishtonpm": "yarn build && yarn lint && yarn test && npm publish"
   },
   "files": [
     "./artifacts/verification_key.json",
@@ -41,6 +35,7 @@
     "buffer": "^6.0.3",
     "circomlibjs": "^0.1.7",
     "json-bigint": "^1.0.0",
+    "pako": "^2.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "snarkjs": "^0.7.2",
