+++ packages/circuits/.gitignore
@@ -2,3 +2,4 @@
 /yarn-error.log
 sha256-var-circom
 /build
+/artifacts
