+++ packages/circuits/build.sh
@@ -8,11 +8,13 @@ PTAU_PATH=$BUILD_DIR/$PTAU
 CIRCUIR_PATH=./circuits/qr_verify.circom
 CIRLIB_PATH=./node_modules
 R1CS_PATH=$BUILD_DIR/qr_verify.r1cs
+QR_VERIFY_DIR=$BUILD_DIR/qr_verify_js
 PARTIAL_ZKEYS_DIR=$BUILD_DIR/partial_zkeys
+ARTIFACTS_DIR=./artifacts
 
 CIRCOM_BIN_DIR=$HOME/.cargo/bin/circom
 
-# install circom and depenencies
+# install circom and dependencies
 function install_deps() {
 
     if [ ! -d $BUILD_DIR ]; then
@@ -20,6 +22,7 @@ function install_deps() {
     fi
 
     echo "Install circom"
+    cd $BUILD_DIR
     if [ ! -f $CIRCOM_BIN_DIR ]; then
         git clone https://github.com/iden3/circom.git
         cd circom
@@ -31,6 +34,7 @@ function install_deps() {
     fi 
 
     echo "Download power of tau...."
+    cd $BUILD_DIR
     if [ ! -f $PTAU_PATH ]; then
         wget https://hermez.s3-eu-west-1.amazonaws.com/$PTAU
         echo "Finished download!"
@@ -60,7 +64,19 @@ function dev_trusted_setup() {
 
     echo "test random" | NODE_OPTIONS='--max-old-space-size=8192' \
 	node ./node_modules/.bin/snarkjs zkey contribute $PARTIAL_ZKEYS_DIR/circuit_0000.zkey $PARTIAL_ZKEYS_DIR/circuit_final.zkey --name="1st Contributor Name" -v 
+    NODE_OPTIONS='--max-old-space-size=8192' ./node_modules/.bin/snarkjs zkey export verificationkey $PARTIAL_ZKEYS_DIR/circuit_final.zkey "$BUILD_DIR"/vkey.json
 
+    echo "Copy proving key and verify key to artifacts!!!!"
+
+        if [ ! -d $ARTIFACTS_DIR ]; then
+        mkdir -p $ARTIFACTS_DIR
+    fi
+
+    cp $QR_VERIFY_DIR/qr_verify.wasm $ARTIFACTS_DIR
+    cp $PARTIAL_ZKEYS_DIR/circuit_final.zkey $ARTIFACTS_DIR
+    cp $BUILD_DIR/vkey.json $ARTIFACTS_DIR
+    
+    echo "Setup finished!"
 }
 
 
@@ -86,15 +102,13 @@ function generate_proof() {
     echo "Building proof...!"
     mkdir -p $BUILD_DIR/proofs
 
-    NODE_OPTIONS='--max-old-space-size=8192' ./node_modules/.bin/snarkjs zkey export verificationkey $PARTIAL_ZKEYS_DIR/circuit_final.zkey "$BUILD_DIR"/vkey.json
-
-    NODE_OPTIONS='--max-old-space-size=8192' ./node_modules/.bin/snarkjs groth16 prove $PARTIAL_ZKEYS_DIR/circuit_final.zkey $BUILD_DIR/witness.wtns $BUILD_DIR/proof.json $BUILD_DIR/public.json
+    NODE_OPTIONS='--max-old-space-size=8192' ./node_modules/.bin/snarkjs groth16 prove $PARTIAL_ZKEYS_DIR/circuit_final.zkey $BUILD_DIR/witness.wtns $BUILD_DIR/proofs/proof.json $BUILD_DIR/proofs/public.json
     echo "Generated proof...!"
 
 }
 
 function verify_proof() {
-    NODE_OPTIONS='--max-old-space-size=8192' ./node_modules/.bin/snarkjs groth16 verify $BUILD_DIR/vkey.json $BUILD_DIR/public.json $BUILD_DIR/proof.json
+    NODE_OPTIONS='--max-old-space-size=8192' ./node_modules/.bin/snarkjs groth16 verify $BUILD_DIR/vkey.json $BUILD_DIR/proofs/public.json $BUILD_DIR/proofs/proof.json
     echo "Verify proof...!"
 }
