+++ packages/circuits/package.json
@@ -7,7 +7,13 @@
     "build": "echo 'no-build'",
     "test": "npx ts-mocha test/**/*.test.ts --exit",
     "format": "prettier --ignore-path .prettierignore --write \"**/*.+(js|ts|json)\"",
-    "lint": "echo 'no lint'"
+    "lint": "echo 'no lint'",
+    "circom": "circom",
+    "dev-install": "/bin/bash ./build.sh install",
+    "dev-setup": "/bin/bash ./build.sh setup",
+    "gen-proof": "/bin/bash ./build.sh gen-proof",
+    "gen-witness": "/bin/bash ./build.sh gen-witness",
+    "verif-proof": "/bin/bash ./build.sh verify-proof"
   },
   "dependencies": {
     "@zk-email/helpers": "^3.0.2",
